chore: Add .gitignore and remove __pycache__
This commit adds a `.gitignore` file to the project to prevent transient Python files from being tracked.

- The `__pycache__` directory is now ignored.
- The previously tracked `__pycache__` directory has been removed.
