+++ quanta_tissu/quanta_tissu/config.py
@@ -1,0 +1,12 @@
+# Toy vocabulary
+vocab = {
+    "hello": 0,
+    "world": 1,
+    "this": 2,
+    "is": 3,
+    "a": 4,
+    "test": 5,
+    ".": 6,
+    "<unk>": 7,
+}
+inv_vocab = {v: k for k, v in vocab.items()}
