+++ quanta_tissu/quanta_tissu/tokenizer.py
@@ -1,0 +1,11 @@
+import numpy as np
+from .config import vocab, inv_vocab
+
+def tokenize(text):
+    tokens = []
+    for word in text.lower().split():
+        tokens.append(vocab.get(word, vocab["<unk>"]))
+    return np.array(tokens)
+
+def detokenize(token_ids):
+    return " ".join(inv_vocab.get(t, "<unk>") for t in token_ids)
